# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central management server for the OpenLabStats fleet. Manages agent enrollment, software mappings, lab groupings, and reporting.",
        "title": "OpenLabStats Server API",
        "contact": {},
        "version": "0.1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/agents": {
            "get": {
                "description": "Returns all agents that have registered with the server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "List all enrolled agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Agent"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/register": {
            "post": {
                "description": "Agents call this on startup and periodically to register themselves with the server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Register or heartbeat an agent",
                "parameters": [
                    {
                        "description": "Agent registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Agent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/rollout-status": {
            "get": {
                "description": "Per-platform rollout progress: how many agents are updated, currently updating, or pending, plus the current version distribution and target.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Agent auto-update rollout status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/{agentID}": {
            "get": {
                "description": "Returns a single agent's details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get agent by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Agent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes an agent from the fleet inventory.",
                "tags": [
                    "agents"
                ],
                "summary": "Remove an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/{agentID}/force-update": {
            "post": {
                "description": "Queues an update for the specified agent which will be delivered on its next heartbeat.",
                "tags": [
                    "agents"
                ],
                "summary": "Force an agent to update immediately",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/{agentID}/lab": {
            "put": {
                "description": "Associates an agent with a lab/room for grouping.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Assign agent to a lab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lab assignment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AssignLabRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/installers/generate": {
            "post": {
                "description": "Returns a download URL for the latest pre-built MSI with an msiexec command.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installers"
                ],
                "summary": "Generate a custom MSI installer",
                "parameters": [
                    {
                        "description": "Installer configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GenerateInstallerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/labs": {
            "get": {
                "description": "Returns all configured labs/rooms.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "List all labs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Lab"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new lab or room grouping for agents.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Create a lab/room",
                "parameters": [
                    {
                        "description": "Lab details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateLabRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.Lab"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/labs/{labID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Get lab by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lab ID",
                        "name": "labID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Lab"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Update a lab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lab ID",
                        "name": "labID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated lab details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateLabRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "labs"
                ],
                "summary": "Delete a lab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lab ID",
                        "name": "labID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/reports/avg-session-time": {
            "get": {
                "description": "Returns top N users ranked by average session duration in minutes. Users with fewer than minLoginsForAvgSession logins in the window are omitted — too small a sample to average meaningfully.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Average session duration per user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to a specific machine",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to a specific lab name",
                        "name": "lab",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/reports/bottom-apps-by-foreground": {
            "get": {
                "description": "Returns bottom (least used) applications by foreground time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Bottom applications by foreground time",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "json",
                        "description": "Output format: json or csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/reports/bottom-apps-by-launches": {
            "get": {
                "description": "Returns bottom (least used) applications by launch count.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Bottom applications by launch count",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "json",
                        "description": "Output format: json or csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/reports/top-apps": {
            "get": {
                "description": "Returns the top applications by total usage seconds over the given time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Top applications by usage time",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "json",
                        "description": "Output format: json or csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/reports/top-apps-by-elevations": {
            "get": {
                "description": "Returns top applications by privilege-elevation launch count over the given time range (UAC on Windows, sudo/admin authorization on macOS). Data is typically sparse.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Top applications by UAC elevation count",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range (e.g. 24h, 7d, 30d)",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to a specific machine",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to a specific lab name",
                        "name": "lab",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "json",
                        "description": "Output format: json or csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/reports/top-apps-by-foreground": {
            "get": {
                "description": "Returns top applications by total foreground (active) time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Top applications by foreground time",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "json",
                        "description": "Output format: json or csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AssignLabRequest": {
            "type": "object",
            "properties": {
                "labId": {
                    "type": "string"
                }
            }
        },
...
        "store.Agent": {
            "type": "object",
            "properties": {
                "agentVersion": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "labId": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "pendingUpdate": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updateOfferedAt": {
                    "description": "Staggered-rollout state. UpdateOfferedAt is nil when the agent is not\ncurrently in flight for an auto-update; UpdateTargetVersion is the version\nit was last offered toward.",
                    "type": "string"
                },
                "updateTargetVersion": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}